location is now double

+++ Scoot_backend/src/main/java/com/example/backend/Entity/Location.java
@@ -5,22 +5,23 @@ import jakarta.persistence.*;
 @Embeddable
 public class Location {
 
-    float longitude;
-    float latitude;
+    double longitude;
+    double latitude;
 
-    public float getLongitude() {
+
+    public double getLongitude() {
         return longitude;
     }
 
-    public void setLongitude(float longitude) {
+    public void setLongitude(double longitude) {
         this.longitude = longitude;
     }
 
-    public float getLatitude() {
+    public double getLatitude() {
         return latitude;
     }
 
-    public void setLatitude(float latitude) {
+    public void setLatitude(double latitude) {
         this.latitude = latitude;
     }
 }
